Create two low level CLIs: Export and Snapshot
